+++ backend/src/nodes/impl/stable_diffusion.py
