+++ backend/src/packages/chaiNNer_stable_diffusion/stable_diffusion/io/load_checkpoint.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import os
 from typing import Tuple
 
-from nodes.impl.stable_diffusion.types import (
+from nodes.impl.stable_diffusion import (
     CLIPModel,
     StableDiffusionModel,
     VAEModel,
