+++ backend/src/packages/chaiNNer_stable_diffusion/stable_diffusion/latent/k_sampler.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import torch
 
 from nodes.group import group
-from nodes.impl.stable_diffusion.types import (
+from nodes.impl.stable_diffusion import (
     Conditioning,
     LatentImage,
     Sampler,
